requirements.txt
asgiref==3.3.1
automium==0.2.6
click==7.1.2
coverage==5.5
Django==3.1.7
django-debug-toolbar==3.2.1
djangorestframework==3.12.4
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
legacy==0.1.6
MarkupSafe==1.1.1
Pillow==8.2.0
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
urlpatterns==0.0.1
Werkzeug==1.0.1
